Highlight nearest tag for coc renderer
Previously, if I had the `:Vista coc` pane open but my cursor was in the code
window, no tag would be highlighted in the vista pane. The `:Vista ctags` pane
would correctly highlight the nearest tag from where my cursor was in the code.

I discovered that this was because the global variables `g:vista.functions` and
`g:vista.raw` are not set when using the coc executive.

This change populates those two globals from list of symbols in the response
from coc. Similar to what happens with the ctags renderer, the `vlnum` values in
`g:vista.raw` are later filled in during rendering the vista window contents.