Add show detail support for nvim_lsp provider
No matter which echo strategy the user choose, as soon as the nvim_lsp
strategy is in use it does not work to show details of the current tag
under cursor. This is due to a missing implementation to extract the tag
of the current Vista buffer line for the nvim_lsp provider. Such has
been added. As result the user can now use the 'p' mapping within the
Vista window, as well as the 'g:vista_echo_cursor' option.